refac: ReviewResult 페이지 리팩토링 - atomic 수정 - 가독성 수정
