+++ src/components/pages/reviewResult/organism/ResultShare.tsx
@@ -1,28 +1,25 @@
 import Button from '@components/common/atom/Button';
-import IconButton from '@components/common/atom/IconButton';
-import SvgIcon from '@components/common/atom/SvgIcon';
 import Typography from '@components/common/atom/Typography';
 import NavigationHeader from '@components/common/molecule/NavigationHeader';
 import FlipCard from '@components/pages/reviewResult/molecule/FlipCard';
-import { Drawer, DrawerContent } from '@components/shadcn/drawer';
+import ShareDrawer from '@components/pages/reviewResult/molecule/ShareDrawer';
 import FixedButtonContainer from '@components/wrapper/FixedButtonContainer';
-import { KakaoShareImage } from '@constants/images';
 import { useFlow } from '@hooks/useStackFlow';
 import { Spacer } from '@nextui-org/react';
 import domtoimage from 'dom-to-image';
 import { saveAs } from 'file-saver';
 import { useRef, useState } from 'react';
-import toast from 'react-hot-toast';
 
 interface ResultShareProps {
   type: string;
   curStep: number;
 }
 
+// TODO 유저네임 변경 필요
 export default function ResultShare({ type, curStep }: ResultShareProps) {
-  const [openBottomSheet, setOpenBottomSheet] = useState<boolean>(false);
-  const ref = useRef<HTMLDivElement>(null);
   const { push } = useFlow();
+  const ref = useRef<HTMLDivElement>(null);
+  const [openBottomSheet, setOpenBottomSheet] = useState<boolean>(false);
 
   const handleClickBackIcon = () =>
     push('ReviewResultPage', { type: type, step: String(curStep - 1) });
@@ -42,20 +39,7 @@ export default function ResultShare({ type, curStep }: ResultShareProps) {
     }
   };
 
-  const handleClickShare = () => {
-    setOpenBottomSheet(true);
-  };
-
-  const handleClickShareLink = async () => {
-    try {
-      await navigator.clipboard.writeText('shareLink');
-      toast.success('링크 복사 완료!', {
-        icon: <SvgIcon id="Complete" color="gray08" />,
-      });
-    } catch (err) {
-      console.log(err);
-    }
-  };
+  const handleClickShare = () => setOpenBottomSheet(true);
 
   return (
     <div className="w-full h-full flex flex-col items-center">
@@ -75,8 +59,11 @@ export default function ResultShare({ type, curStep }: ResultShareProps) {
           textAlign: 'center',
         }}
       />
+
       <Spacer y={10} />
+
       <FlipCard ref={ref} />
+
       <FixedButtonContainer direction="row">
         <Button buttonVariant="tertiary" onClick={handleClickDownload}>
           <Typography variant="body01">카드 저장하기</Typography>
@@ -87,37 +74,11 @@ export default function ResultShare({ type, curStep }: ResultShareProps) {
           </Typography>
         </Button>
       </FixedButtonContainer>
-      <Drawer
-        open={openBottomSheet}
-        onOpenChange={open => setOpenBottomSheet(open)}
-      >
-        <DrawerContent className="mx-auto max-w-[600px]">
-          <div className="p-4 flex justify-center gap-4">
-            <div className="flex flex-col gap-2 items-center">
-              <IconButton
-                iconProps={{
-                  id: 'Share',
-                  width: 36,
-                  height: 36,
-                  color: 'primary',
-                }}
-                onClick={handleClickShareLink}
-              />
-              <Typography variant="body03" className="text-center">
-                링크복사
-              </Typography>
-            </div>
-            <div className="flex flex-col gap-2 items-center">
-              <button className="h-[36px]">
-                <img src={KakaoShareImage} style={{ width: 36, height: 36 }} />
-              </button>
-              <Typography variant="body03" className="text-center">
-                카카오톡
-              </Typography>
-            </div>
-          </div>
-        </DrawerContent>
-      </Drawer>
+
+      <ShareDrawer
+        openBottomSheet={openBottomSheet}
+        setOpenBottomSheet={setOpenBottomSheet}
+      />
     </div>
   );
 }
